added dbgtypes.cpp/dbgtypes.hpp, and started to cleanup main dbg class
