+++ .gitignore
@@ -1,0 +1,7 @@
+build
+build/*
+dbg
+disass
+elf
+
+
