+++ Makefile
@@ -3,7 +3,7 @@ LIBNAME = capstone
 TESTLIB = gtest
 FLAGS = -Wall -std=c++17
 
-OBJECTS = build/dbg.o build/elf.o build/elftypes.o build/disass.o
+OBJECTS = build/dbg.o build/dbgtypes.o build/elf.o build/elftypes.o build/disass.o
 ELF_OBJECTS = build/elf.o build/elftypes.o build/disass.o
 DISASS_OBJECTS = build/disass.o
