+++ dbg.cpp
@@ -14,122 +14,7 @@
 #include <map>
 #include "elftypes.hpp"
 #include "elf.hpp"
-
-enum disas_mode {
-  DISAS_MODE_WORD,
-  DISAS_MODE_BYTE
-};
-
-struct command {
-  std::string cmd;
-  std::vector<std::string> args;
-} typedef command_t;
-
-class Breakpoint {
-private:
-    uint64_t addr;
-    uint8_t data;
-    bool active; 
-
-public:
-  Breakpoint(uint64_t bp_addr, uint8_t orig_data) : addr(bp_addr), data(orig_data), active(true)
-  {}
-  
-  uint64_t get_addr() const {
-    return addr;
-  } 
-  
-  uint64_t get_data() const {
-    return data;
-  }
-
-  uint64_t get_mod_data() const {
-    unsigned long new_data = ((data & ~0xff) | 0xcc);
-    return new_data;
-  }
-
-  void enable() {
-    active = true;
-  }
-
-  void disable() {
-    active = false;
-  }
-};
-
-
-class MapEntry {
-private:
-  uint64_t start_addr; 
-  uint64_t end_addr; 
-  bool permissions[4]; // R W X P
-  uint32_t size;
-  uint32_t offset;
-  std::string file;
-  std::string permissions_str;
-
-public:
-  MapEntry(std::string entry_str) {
-    std::string start_addr_str;
-    std::string end_addr_str;
-    std::string offset_str;
-
-    std::stringstream entry(entry_str);
-
-    std::getline(entry, start_addr_str, '-');
-    std::getline(entry, end_addr_str, ' ');
-    std::getline(entry, permissions_str, ' ');
-    std::getline(entry, offset_str, ' ');
-
-    while (getline(entry, file, ' '));
-     
-    start_addr = std::stol(start_addr_str, NULL, 16); 
-    end_addr = std::stol(end_addr_str, NULL, 16); 
-
-    offset = std::stol(offset_str, NULL, 16);
-    if (permissions_str.size() != 4) {
-      std::cout << "could not get permissions" << std::endl;
-      return;
-    }
-
-    for (int i = 0; i < 4; i++) {
-      if(permissions_str[i] != '-') {
-        permissions[i] = true;
-      } else {
-        permissions[i] = false;
-      }
-    }
-
-    size = end_addr - start_addr;
-  }
-
-  uint64_t get_start() {
-    return start_addr;
-  }
-
-  uint64_t get_end() {
-    return end_addr;
-  }
-
-  uint32_t get_size() {
-    return size;
-  }
-
-  bool contains(uint64_t addr) {
-    if (addr >= start_addr && addr < end_addr) {
-      return true;
-    }
-    return false;
-  }
-
-  std::string str() {
-    std::stringstream ss;
-    ss << "0x" << std::hex << start_addr << "-0x" << std::hex << end_addr << "   " << permissions_str << "      " 
-       << std::hex << size << "  " << std::hex << "   " << offset <<"   "<< file;
-
-    return ss.str();
-  }
-};
+#include "dbgtypes.hpp"
 
 class Debugger {
 private:
@@ -208,7 +93,10 @@ private:
     
     for (std::string line; getline(vmmap_file, line); ) {
       MapEntry map(line);
-      //std::cout << map.str() << std::endl;
+      if (map.get_start() == 0) {
+        continue;
+      }
+
       vmmap.emplace_back(map);
     }
   }
